Better component compartmentalization for app market
